crud cliente

+++ src/main/java/web/dao/ClienteDao.java
@@ -40,7 +40,7 @@ public class ClienteDao {
 		return manager.find(Cliente.class, id);
 	}
 
-	public void remove(int id) {
+	public void remove(long id) {
 		manager.createQuery("delete from Cliente c where c.id = :id").setParameter("id", id).executeUpdate();
 	}
